SCRIPTED-28 Mark occurrences
Fully implemented now.  Configuration options are available:

{
	mark_occurrences : {
		interval :500,
		disable : false,
		retain : false
	}
}

interval is the amount of time between moving the caret and invoking the mark_occurrences operation.
disable if set to true will completely disable mark occurrences
retain if set to true will retain the previous marks if the current caret position is not over a word.