Adding content assist support for Closure
It doesn't look like there's another mechanism to add support for
content assist other than adding files directly into server code.
As a result, this PR is somewhat invasive.

The way `goog.require` and `goog.provide` works is so that when the
code is put through the closure compiler, it all gets stripped away
and the files are all located in the proper order. In development
mode, you can use the
[DepsWriter](https://developers.google.com/closure/library/docs/depswriter)
to generate an intermediate file that builds a dependency tree.
This PR relies on that intermediate file to locate the correct files
(in the resolver) for content assist. Here are the added functionalities:

- Support for `.scripted` configuration to point to the deps file:
    `{ "closure" : { "deps" : "relative/path/to/deps.js" } }`
- A deps file loader (regex) to map modules/constructors to file
- Adds a new `kind` of module, `"closure"`.
- Treats the `"closure"` type as global (namespaced of course)